manager: fix a lot of problem
Todo:
More migrate
e.g description card
Cleanup strings
