+++ manager/app/src/main/res/values-zh-rCN/strings.xml
@@ -467,11 +467,11 @@
     <string name="static_kstat_config">静态 Kstat 配置</string>
     <string name="kstat_path_management">Kstat 路径管理</string>
     <string name="no_kstat_config_message">暂无 Kstat 配置，点击下方按钮添加配置</string>
+    <string name="susfs_basic_information">基本信息</string>
     <string name="susfs_hide_mounts_control_title">SuS 挂载隐藏控制</string>
     <string name="susfs_hide_mounts_control_description">控制 SuS 挂载对进程的隐藏行为</string>
-    <string name="susfs_hide_mounts_for_all_procs_label">对所有进程隐藏 SuS 挂载</string>
-    <string name="susfs_hide_mounts_for_all_procs_enabled_description">启用后，SuS 挂载将对所有进程隐藏，包括 KSU 进程</string>
-    <string name="susfs_hide_mounts_for_all_procs_disabled_description">禁用后，SuS 挂载仅对非 KSU 进程隐藏，KSU 进程可以看到挂载</string>
+    <string name="susfs_hide_mounts_for_nonsu_procs">对非 Root 进程隐藏 SuS 挂载</string>
+    <string name="susfs_hide_mounts_for_nonsu_procs_description">决定是否要对非 Root 进程隐藏 SuS 挂载</string>
     <string name="susfs_hide_mounts_all_enabled">已启用对所有进程隐藏 SuS 挂载</string>
     <string name="susfs_hide_mounts_all_disabled">已禁用对所有进程隐藏 SuS 挂载</string>
     <string name="susfs_hide_mounts_recommendation">建议在屏幕解锁后或在 service.sh 或 boot-completed.sh 阶段设置为禁用，这可以修复一些依赖 KSU 进程挂载的 root 应用的问题</string>
