+++ manager/app/src/main/res/values-zh-rTW/strings.xml
@@ -472,9 +472,6 @@
     <string name="no_kstat_config_message">無 Kstat 設定，點選下方按鈕新增設定</string>
     <string name="susfs_hide_mounts_control_title">SuS 掛載隱藏控制</string>
     <string name="susfs_hide_mounts_control_description">控制 SuS 掛載對處理程序的隱藏行為</string>
-    <string name="susfs_hide_mounts_for_all_procs_label">對所有處理程序隱藏 SuS 掛載</string>
-    <string name="susfs_hide_mounts_for_all_procs_enabled_description">啟用後，SuS 掛載將對所有處理程序隱藏，包含 KSU 處理程序</string>
-    <string name="susfs_hide_mounts_for_all_procs_disabled_description">停用後，SuS 掛載只對不是 KSU 處理程序隱藏，KSU 處理程序將可以看到掛載</string>
     <string name="susfs_hide_mounts_all_enabled">已啟用對所有處理程序隱藏 SuS 掛載</string>
     <string name="susfs_hide_mounts_all_disabled">已停用對所有處理程序隱藏 SuS 掛載</string>
     <string name="susfs_hide_mounts_recommendation">建議在 service.sh 或 boot-completed.sh 階段，或是在螢幕解鎖後設為停用。這可以修復一些依賴 KSU 處理程序掛載 root 應用程式的問題</string>
